+++ app/main.py
@@ -1,5 +1,10 @@
+from src.logger.logging import setup_logging
+from src.tools.utils import parse_arguments
+
+
 def main():
-    pass
+    args = parse_arguments()
+    setup_logging(log_level=args.log_level, log_level_others=args.log_level_other)
 
 
 if __name__ == "__main__":