+++ CHANGELOG.md
@@ -10,3 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Planned
 - Use SQLite DB for storing data
+
+## 2026-02-11
+### Added
+- Logging
+- Argument Parser
+- Isort and Black configuration
+- Loading environment values